requirements.txt
django-cassandra-common
cassandra-driver==3.23.0
Django==3.0.7
django-cassandra-engine==1.6.1
djangorestframework==3.11.0
django-filter==2.3.0
drf-yasg==1.17.1
requests==2.23.0
django-cors-headers==3.4.0
django-environ==0.4.5